requirements.txt
streamlit
requests
fastapi
uvicorn
python-dotenv
google-generativeai
